plainbox:secure:plugins: add IPlugInCollection.{discovery_time,get_total_time}
This patch adds one property and one method on the plug-in collection
interface.

The discovery_time property describes the amount of time that was spent
while the collection was looking for plugins. This typically reflects
file system traversal and filtering time.

The get_total_time() method aggregates discovery time with load and wrap
time of each plugin to give the total cost of using a given collection.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>