plainbox:secure:plugins: add IPlugIn.plugin_{load,wrap}_time properties
This patch adds two new properties to the IPlugIn interface, that can be
used to inspect the overhead incurred by loading and wrapping
(processing) a given plugin (loadable entity).

In addition, both properties are implemented on all the concrete
sub-classes of the interface, namely the PlugIn, UnitPlugIn,
WhiteListPlugIn and Provider1PlugIn classes.

Since the load time is external (i.e. the collection loads the plugin,
not the plugin itself) all IPlugIn sub-classes gained a new, mandatory
argument to __init__(), which holds the load time. Subsequently all of
the collection classes FsPlugInCollection and
PkgResourcePlugInCollection were patched to measure that time and pass
it along.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>