plainbox:session:resume: track session directory
This patch makes SessionResumeHelper track the location of the session
directyry. The location is taken from the SessionStorage object and
passed to the resume helper. This allows the resume helper to optionally
rewrite the location of each log file so that it is relative to the
actual location of the session file.