plainbox:session:resume: add 'rewrite-log-pathnames' flag
This patch adds a new resume flag FLAG_REWRITE_LOG_PATHNAMES, that can
be used to control if the session log files are to be re-written if they
cannot be found at their nominal location. This can often happen while
moving session from machine to machine.

Note that to use this feature, the FLAG_FILE_REFERENCE_CHECKS flag must
be set as well, otherwise broken references are not detected and rewrite
logic is not active.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>