plainbox:vendor:extcmd: add on_abnormal_end() to delegates
This patch adds a new method that all delegates must support,
on_abnormal_end, which signals that a process was killed by a signal.
This is distinct from the process just exiting with a failure code.

The new method is added to the IDelegate interface, as well as to each
concrete class.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>