checkbox-touch: set effective category identifiers
This pach mirros the previous patch to plainbox and makes checkbox touch
set the effective category identifier every time a test plan is
selected. This patch is not fully correct as it doesn't handle generated
jobs (and doesn't keep the effective category id updated for jobs that
got added during testing) but so far those are not supported by checkbox
touch much.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>