providers:checkbox:bin:pm_test: Ensure that ~/.config/autostart is not owned by root
pm_test has to be run by the root user, by using pkexec or sudo.
Problem is that if ~/.config/autostart does not exists the script did make
it without chown it back to the normal user.

The fix just calls os.chown() on both .config and autostart with normal uid/gid

Fixes: https://bugs.launchpad.net/plainbox-provider-checkbox/+bug/1382321