plainbox:units:job: add qml plugin type
This patch adds a new plugin type 'qml' that will be used for native QML
test jobs. Such jobs will not rely on a test command but instead will
require applications to load a QML component into a test shell (either
the application itself or a stadalone test shell) that exposes
appropriate APIs for test developers.

Test developers can implement any desired logic in the QML (and
typically javascript), having access to many platform APIs, and then
report back to the test shell with the outcome of the test.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>