plainbox:secure:plugin: add the LazyPlugInCollection
This patch adds a new plug in collection class, LazyPlugInCollection,
which is conceptually similar to the PkgConfigPlugInCollection but has
different performance characteristics. Each element can be loaded from a
pkgconfig-like string ('package.module:symbol'). Elements can be loaded
all at once, or lazily one at a time. The important property, not shared
with pkgconfig, is that one can discover all the keys efficiently. This
allows it to be used as an optimization for many cases that we currently
use plugins with.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>