plainbox:session: respect PLAINBOX_SESSION_REPOSITORY
This patch changes SessionStorageRepository.get_default_location() to be
a more flexible and reliable.

First of all, it now respects the new environment variable,
PLAINBOX_SESSION_REPOSITORY. This variable overrides the default
location of $XDG_CACHE_HOME/plainbox/sessions.

The effective repository location is sanitized so that pointing it to a
file or to a read-only location won't be a fatal problem anymore.
Instead a sensible error message is logged and a emergency, temporary,
session repository is created.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>