plainbox:secure:plugins: override LazyPlugInCollection.discovery_time
This patch allows to access the discovery_time property on a
LazyPlugInCollection before discovery is complete.

This is only natural as the lazy plugin collection is often used in a
lazy (non-complete, on-demand) loading scenarios. The problem is that
generic parts of plugin handling code might access the discovery_time
property (for reporting) and would get an AttributeError before
discovery was complete.

This patch removes the check and simply returns zero before discovery is
really complete.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>