plainbox:excmd: don't expose non-blocking pipes to children
This patch fixes a bug in the glibc-based extcmd implementation.

The communication pipes that are created to pass stdout and stderr data
were created in non-blocking mode and that non-blocking pipe was exposed
to the child process. A child process writing fast enough could fill the
pipe and get EAGAIN. Since applications are not generally designed to
handle that it would cause random failures.

The fix based on creating normal (blocking) pipes and only set the
non-blocking mode in the parent process that has an event loop.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1401886

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>