plainbox:vendor:extcmd: don't block SIG{INT,QUIT,CHLD} in test processes
This patch corrects a bad assumption that the mask of blocked signals is
reset across exec. It's not, thus to let the test process be killable
with SIGINT or SIGQUIT (and more importantly, to let that process manage
its own SIGCHLD) we must remove the mask that was applied in the parent.

In practical terms we can now CTRL-C or CTRL-\ to interrupt / quit the
started test process.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1402680