checkbox-gui:gui-engine: Only use the current job metadata when resuming a session
The fix for bug https://bugs.launchpad.net/bugs/1362449 uses the current job
got from the session metadata when resuming checkbox-gui.

Starting a fresh session works as expected only if all the jobs are selected
because in that case the current job metadata is empty. But if the tester
only selects a subset of the whitelist or just one test it fails and ends the
testrun with no results.

So we should only use the metadata when resuming a session and still use
NextRunJobIndex(-1) in other cases.