checkbox-gui:gui-engine: Increment the current job_index when forcing the outcome
When resuming a session, the tester can force the outcome of the previous test
from the resume screen.
Since the fix for https://bugs.launchpad.net/bugs/1362449 the current job is
extracted from plainbox metadata.
But we have to increment the current index when forcing the test outcome to
automatically go to the next job in the run list (or complete the testrun if it
was the last one).
Similar to the previous bugfix, a new boolean "m_resuming_force_result"
stores the tester decision and help decide if we have to increment the current
index.

Fixes https://bugs.launchpad.net/checkbox-gui/+bug/1406719