plainbox:ctrl: pass job_state inside execution controllers
This patch pushes the job_state object through the execution controller
call stack so that at the one point where we currently look at job.via,
we can, instead, look at job_state.via_job.

This patch has one annoying bug that is highlighted in the patch, it
passes job_state of a generated job to the method that deals with the
generator. It probably means that the patch is incorrect and we need to
pass more job_state objects (ugh!) or to pass the state_map directly
(ugh too).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>