plainbox:job: rewrite JobTreeNode construction code
This patch rewrites the construction code of JobTreeNode (aka
JobTreeNode.create_tree()) to use the new helper class, TreeBuilder.

The builder contains local state and support methods to effectively do
the same thing, that is, to build a tree of nodes that represent
categories and jobs. Not-so-intuitively, each node can contain a
"label", any set of "jobs" (leaves) and "categories" (nodes).

On the outside the only difference is a slight API change to
JobTreeNode.create_tree() that now requires the session_state argument
which is needed to look at mutable job state.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>