providers:checkbox: run {poweroff,reboot} in the foreground
This patch changes the pm_test script to run either poweroff(8) or
reboot(8) commands in the foreground and then wait indefinitely until
something kills the pm_test process.

This prevents LP: #1413132 from occurring by letting plainbox wait until
the script finishes (which it never does). This is a stop-gap measure
until noreturn flag is implemented and understood by plainbox and all
appropriate jobs are patched to use it.

Fixes: https://bugs.launchpad.net/plainbox-provider-checkbox/+bug/1413134
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>