plainbox:session: move JobReadinessInhibitor checks after initialization
This patch moves all JobReadinessInhibitor.__init__() checks to happen
after the super-call to init. Instead of checking function arguments
they now check instance attributes. This sets the stage for later
changes that removes some of the (redundant) checks and alter the rest.

NOTE: The check that changes cause *no longer* assigns it back as cause
is a read-only field.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>