plainbox: combine IProvider1 and IProviderBackend1
This patch combines the two interface classes into one provider interface.
In the past it was designed to separate the "backend" code from
the "frontend" code so that mocking could be used to emulate just one
side of a provider. In practice this seemed to have little value so it's
easier to think of just one provider interface.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>