plainbox:secure:providers: add new provider loader support classes
This patch adds the following new clasess:
 - ProviderContentPlugIn
 - WhiteListPlugIn2
 - UnitPlugIn2
 - ProviderContentEnumerator
 - ProviderContentClassifier
 - ProviderContentLoader

Together they can be used to seed a provider object with actual content.
The enumerator is used to locate all of the files needed to load
the provider. It can be easily mocked to totally replace what a provider
is without touching any other code. The classifier can determine
the purpose of each file relevant to a working provider. Together those
classes are used by the loader, which uses the classifier to delegate
loading each file to a particular ProviderContentPlugIn sub-class.

The three content plug-in classes can load:
 ProviderContentPlugIn -> convert any file to a FileUnit
 UnitPlugIn2 -> convert any .pxu file to a list of Unit objects,
                convert any test plan units into WhiteList objects
 WhiteListPlugIn2 -> convert any .whitelist file into a WhiteList
                     object and the corresponding TestPlanUnit

Some classifiers don't return a plugin loader class to indicate that
we don't really care about units of that particular kind. For example,
we don't create a FileUnit for each file that is a part of a version
control system, because we have no need for that.

NOTE: Both WhiteListPlugIn2 will become WhiteListPlugIn, UnitPlugIn2
will become UnitPlugIn in one of the subsequent patches.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>