plainbox:vendor: vendorize morris 1.2
This patch adds morris 1.2 (https://pypi.python.org/pypi/morris) into
the plainbox vendor module. Morris is the evolution of the code that
started as the ``plainbox.impl.signal``. Morris is licensed under LGPLv3
(same as the code it replaces) and is Copyright (c) 2012-2015 Canonical.

In addition, the same patch changes morris so that all imports are from
plainbox.vendor.morris. This ensures that if a different or incompatible
version of morris is installed globally it will never conflict with our
copy.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>