plainbox:units:testplan: use xscanners to handle category overrides
This patch removes the hand-crafted shlex.split-based parser (that never
really worked correctly but we never tried it to that extent) and replaces
it with the new xparsers-based code. While the scanner underneath
is nothing I am proud of, it gets the job done.

The code is designed to inspect a OverrideList node which is composed of
a list of OverrideField or Error nodes. Error nodes cause an exception
to be raised (as it was in the original implementation). OverrideFields
are inspected and both the category_id and the selection pattern
are qualified according to the namespace of the provider.

There are undoubtedly differences between how this code and the one before
it operate but I think the change is only for the better.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>