plainbox:unit:testplan: add TestPlanUnitSupport
This patch adds TestPLanUnitSupport, a new support class that
hopefully, begins the theme of support companion classes and
more-or-less declarative unit classes.

This particular class duplicates the existing functionality of a
TestPlanUnit to compute the effective qualifier and also adds entirely
new functionality to compute the list of overrides stored in a test
plan.

The intent is to later on replace specific parts of TestPlanUnit class
with simple pass-through calls to the support class and keep all of the
more complicated logic away from the unit-is-a-set-of-fields code.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>