plainbox:pod: add unique_elements_assign_filter() aka unique()
This patch adds a new assign filter that ensures that all elements of
the sequence that is being assigned to a field are unique.

This feature is useful when coupled with tuples (which are immutable)
and change notification. It allows applications to make modifications to
a sequence of objects by assigning new tuples (so "editing" the sequence
is possible) while retaining the simple check that ensures that no
accidental edits cause duplicates.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>