plainbox:unit:testplan: give less incorrect advice for pattern validation
This patch builds upon the earlier realization that select_jobs() was
buggy and refrains from telling the user in strong words that a test
plan (or whitelist) is totally broken since we cannot determine that
reliably yet.

For all the cases where we don't know for sure that a given entry is
correct we emit an advice to say that it might be incorrect. This is
better from UI point of view but also fixes an issue where a correct
test plan (such as the one described in the commit message of the fix to
select_jobs() earlier) is incorrectly classified as buggy.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>