plainbox:xparsers: add WordList parser
This patch adds a small new parser for a simple list of words. The list
follows the checkbox tradition of being permissive, allowing both
white-space and commas as separators (and in fact, it also allows
seemingly erroneous separator chains).

The primary use case of this word list so far is to parse the requires
and depends fields, so that jobs can use quoted template expansions and
still be safe, that nothing weird is going to happen there.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>