providers:checkbox: fix staging/disk/stats_{name}
This patch corrects the prior correction to the
staging/disk/stats_{name} template. The original intent was to prevent
instantiation of this template from devices that lack the ``name``
attribute. This cannot be done by actually using the {name} expansion in
the requires: field as that field is a part of the expansion. Instead,
the ``template-filter`` can be used to require a non-empty ``name``.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>