plainbox:provider_manager: add minimal metadata to job tests
This patch adds minimal meta-data to all the test jobs used internally
by the provider manager module. This includes the description and the
estimated duration. Oddly it doesn't include the summary. I suspect it's
a bug where summary defaults to partial_id but it also validation gets
confused and doesn't report it. I didn't want to fix it in this commit
so I left summary out.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>