plainbox: implicitly load the manifest provider
This patch changes how plainbox implements the ``all_providers`` for the
base "checkbox-like" tool command mix-in (which some, but not all, tools
use). From now on it automatically include the special provider
``manifest`` and has provisions to also include the ``category``
provider.

This work requires more research and will finish with a CEP proposal
about core / special providers that are automatically present in all
applications. For now the manifest provider is trying out new grounds.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>