providers:checkbox: add monitor/thunderbolt job
This patch adds a clone of the 'monitor/hdmi' job for testing
Thunderbolt displays. The test depends on the manifest resource,
specifically on the 'has_thunderbolt' manifest entry.

I've stripped out the wording that said to skip the test if there is no
such port since that's what the manifest is for to tell us.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>