plainbox:job: fix potential infinite recursion
This patch fixes a potential infinite recursion that may happen due to a
bug in sequence of calls in JobTreeBuilder.get_or_create_category_node().

The builder will create a parent category for each category that gets
added but will only update the category map once the whole operation is
complete. Since the function calls itself recursively it can keep trying
to create one category node over and over forever.

This patch moves the code that updates category_node_map directly to
JobTreeBuilder.create_category_node() so that the map is always
up-to-date.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1442148

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>