plainbox:vendor:extcmd: use 1-byte buffers for legacy chunked IO
This patch is a hack to get to an useful effect from legacy (thread
based) chunked IO implementation. By default chunked mode used a 4K
buffer which was a reasonable value for bulk data transfers. Since the
blocking API doesn't allow us to see anything before the whole buffer
gets full we cannot use it to get things like interactive prompts to
work.

In order to fix bug https://bugs.launchpad.net/plainbox/+bug/1443911 I
wanted to use chunked IO (and it works great in glibc mode which uses
non-blocking IO). To use chunked IO in thread + blocking mode we can
cheat and use 1-byte buffers. That is, each byte that gets written to a
pipe gets read and reported to the application. This is slower and by
far, but it gets the job done.

Since chunked IO is enabled selectively and so far only the manifest job
enables it, it's a safe hack / compromise.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>