plainbox:commands:run: flush on each part of IO
This patch enables explicit flush on each write that came from a job
process. This is normally line-based IO for all the text-based jobs.

When using CHUNKED_IO with legacy thread+blocking IO extcmd this is
actually triggered on each byte read so we get "live" output, as it
happens.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>