plainbox:depmgr: don't crash when directly solving unknown jobs
This patch changes the DependencyManager class not to crash when asked
to solve the dependency chain of a given job, directly (as if it was
placed on the desired_job_list) but the job is not present in the solver
copy of job_list.

Instead of raising KeyError, a new error (subclass of DependencyError)
DependencyUnknownError is raised. This error is handled uniformly by the
SessionState class so everything works as expected.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1334296

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>