checkbox-touch: add ListItemWrappable component
This patch adds a list item class that handles multiline labels more gracfully.
It's a subclass of ListItems.Empty with the standard bits 'borrowed' from
ListItems.Standard. It's not a subclass of the Standard list item, as it
internalises some bits that this class have to fiddle with.
By having this class it's also possible to fully tweak look and feel of the
lsit item as checkbox-touch evolves.
This component should serve as a replacement to all ListItem.Standard uses in
the code.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>