checkbox-touch: fix unlatching on selections
This patch fixes when buttons got unlatched on selection screens.
Previously unlatching was called as soon as call to python was done (for
instance when remembering categories), which in fact made the application
unlatch buttons before back-end processing (python side) was done. This patch
calls unlatching in the follow-up of the python method call making unlatching
the last action.

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>