checkbox-touch:ap-tests: make skip_test wait for an object
This patch makes skip_test helper function wait for a given object to become
selectable and visible before skipping action is performed. This fixes a
problem when due to a race condition, click_action was called on an empty
header (empty, because it hasn't been yet populated by the page being brought
up).

Fixes: https://bugs.launchpad.net/checkbox-touch/+bug/1450165
Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>