checkbox-gui: Add the DarkDialog qml component
In order to still get the dark background popup that we used to gave with the
first version of the Ubuntu SDK, the DarkDialog component now provides the
Dialog popups for checkbox-gui.

Overriding StyledItem is not possible so the Main Container is an UbuntuShape
instead of a FocusScope based element.