checkbox-support:parsers:submission: Add parsing of modprobe_attachment.
This includes the simple parser as well as the changes to
SubmissionParser and the TestRun class to actually invoke the subparser
when needed and gather the data.

Testing changes add a unit test for the parser,  the xml chunk data file
for the test, the test method itself, and augment the SubmissionRun
data-holding class with a addModprobeInfo method which we'd expect the
submission parser to call when it finds and parses a
modprobe_attachment.