checkbox-support:parsers:submission: Add parsing of PCI subsystem ID.
This is simply obtaining the subsystem ID from the first device as output
by lspci -x (which we have in the lspci_standard_config_attachment). This is
used to uniquely identify some systems.

This includes the simple parser as well as the changes to SubmissionParser and
the TestRun class to actually invoke the subparser when needed and gather the
data.

Also includes unit tests for the parser, integration test for the
SubmissionParser with XML data file, and an entry point so the parser can be
invoked with plainbox dev parse.