plainbox:pod: add unset variants for type assignment filters
This patch adds pod.unset_or_typed and pod.unset_or_typed.sequence
assign filters. Those two new filters allow to use UNSET values (when
explicitly desired) and still get a benefit from the type checks that
typed and typed.sequence offer.

UNSET acts as a kind of variant of None, offering nullable types. This
is not the best way to do it but I wanted to have a distinction between
None and UNSET so that None can be used normally.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>