plainbox:pod: filter-out UNSET from as_dict()
Normally UNSET values don't exist in a POD for a long time. Typically
they are only used briefly during POD initialization. If the developer
uses a custom default value and actually uses UNSET (or assigns UNSET
explicitly) then as_dict() will misbehave, producing the UNSET value
again (which is implicit).

This patch changes as_dict() so that it doesn't propagate them to the
dict representation. Note that this is different from the as_tuple(),
which cannot avoid representing UNSET.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>