plainbox:providers:special: allow passing keyword args to get_stubbox()
The provider constructor seems to call the old validate API. Since this
is pretty annoying (all the incorrect things the old validation says) I
wanted to have a way to disable that by passing validate=False,
check=True. This is a quick stop-gap measure. In reality we want to
purge all the old validation code.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>