checkbox-ng:dbus_ex: add setter for ObjectWrapper.native
This patch makes it possible to re-target an ObjectWrapper so that from
the outside, nothing changes, but on the inside, the object being
wrapped (the python object being wrapped with a dbus interface) is
indeed changed.

The main motivation for this change is to make it possible to re-wire
results on any change (which are about to become immutable).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>