checkbox-support:lib:dmi: Add access to "raw" attributes from DMI.
The DmiDevice is initialized with a dictionary of data items obtained from DMI.
Normally they can be accessed only via a specific set of getters. Rather than
augment the list of getters which can become cumbersome, this method allows optional
access to the "raw" data in the initializing dictionary. This should only be used
if either the raw data is really desired, or if there's no accessor for the data item,
because this bypasses the heuristics that some of the existing accessors apply.