checkbox-support:parsers:submission: Add "raw" DMI device support.
This leverages the DMI parsing enhancements to give special treatment
to BIOS and SYSTEM devices. If one of those is found, in addition to
the existing "addDmiDevice" call which adds a device (which always
"normalizes" the set of attributes into the ones that make sense for an
actual device, thus discarding some DMI-specific data), a new
addRawDmiDevice method is called on the result. This way the result
class is free to inspect and store a DMI device's raw data.