providers:qml: add a simple automated battery test
This patch adds a simple test for battery.
If more than 5% of total battry capacity is lost in 30m then it would be
considered a fail.

When the screen is locked, all the Timers that are defined in the qml tests
are frozen, making such test useless if nothing can wake up the application.

Registering an Alarm (the duration of the battery test) ensures that the battery
percentage will be refreshed if the screen is unlocked.