plainbox:session:manager: add SessionManager.create_exporter()
This patch adds a simple method to create an exporter with the specified
ID and set of options. This is meant to simplify all the places that need
to do this but need to poke the lower level APIs to achieve that.

You may note an inconsistency in the argument name, the exporter_support
object is passed as exporter_unit. This is deliberate. The exporter_unit
argument name is wrong but fixing this spills to subsequent commits where
all the code is updated to use this API.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>