plainbox:command:run: don't inspect result builder directly
This patch fixes an issue where the job runner UI would misbehave and show
weird UNSET values (and lose some of the former feature set).  This was
caused by the introduction of the job builder. The defaults on a builder
object are different than those on a result object. In result (sic) logic
that triggered on None defaults stopped working.

This patch fixes it by changing the affected code to read a temporary
result object created from the builder. This way all writes are done to
the builder and all reads are done from a result copy.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1476136
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>