plainbox:ingredients: add PlainboxCommand and friends
This patch adds a set of guacamole extensions (ingredients) that make
use of plainbox from guacamole-based applications easier. This constitutes
the CanonicalCrashIngredient, RenderingContext
and SessionAsisstantIngredient.

The RenderingContextIngredient adds the ``rc``attribute to the guacamole
context. This represents the RenderingContext which is a fancy name for
a stateful text paragraph rendering API.

The CanonicalCrashIngredient replaces the stock crash ingredient and
is responsible for displaying a more visually pleasant message along with
simple instructions on how to report the problem.

The SessionAssistantIngredient adds the ``sa`` attribute to the guacamole
context. This represents the SessionAssistant. A new API, still somewhat
under construction, that is meant for simple python scripts, as away
to execute non-interactive tests with more control than what is granted
by the launcher infrastructure in checkbox.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>